requirements.txt
pymongo>=4.6
python-dotenv>=1.0
fastapi>=0.111
uvicorn>=0.30
pydantic>=2.7
